Lots of small tweaks, runs okay - not every script is setup. A nice template with README updates for anyone thats wants to clone

+++ bin/docker.sh
@@ -1,12 +1,18 @@
 #!/bin/bash
+# Variables/Logging
+source $PWD/bin/_exports.sh
+FILE=$(basename "$0")
+
 if [[ ! $INSTALL_SCRIPT ]]; then
-  echo "(!) Error: You must use the installer script."
-  exit
+  error "Error: You must use the installer script."; exit
 fi
 
-COMPOSE_VERSION="1.23.2"
+# Argument Required for File
+if [ -z "$1" ]; then
+  error "Developer Error! Missing Argument for $FILE"; exit
+fi
 
-echo "(+) Installing Docker CE | First Removing Old Docker Engine"
+info "(+) Installing Docker CE | First Removing Old Docker Engine"
 sudo apt-get remove -y docker docker-engine docker.io containerd runc
 
 sudo apt-get -y install \
@@ -15,30 +21,30 @@ sudo apt-get -y install \
   gnupg \
   lsb-release
 
-echo "(+) Installing Docker GPG Key"
+info "(+) Installing Docker GPG Key"
 sudo mkdir -p /etc/apt/keyrings
 curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo gpg --dearmor -o /etc/apt/keyrings/docker.gpg
 
-echo "(+) Updating the /etc/sources.list.d/"
-echo \
+info "(+) Updating the /etc/sources.list.d/"
+info \
   "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.gpg] https://download.docker.com/linux/ubuntu \
   $(lsb_release -cs) stable" | sudo tee /etc/apt/sources.list.d/docker.list >/dev/null
 
 apt-get update
 
-echo "(+) Installing Docker"
+info "(+) Installing Docker"
 sudo apt-get install -y \
   docker-ce \
   docker-ce-cli \
   containerd.io \
   docker-compose-plugin
 
-echo "(+) Adding user to docker group"
+info "(+) Adding user to docker group"
 # Create a docker group, add our user to it
 sudo groupadd docker
-sudo usermod -aG docker $USER
+sudo usermod -aG docker "$USER"
 sudo usermod -aG docker root
 
-echo "(+) Complete! Run with $ docker / $ docker-compose"
-echo -e "\t You may need to reload your shell to apply the new user group permissions."
-echo -e "\t Which were: sudo usermod -aG docker \$USER"
+success "(+) Complete! Run with $ docker / $ docker-compose"
+success -e "\t You may need to reload your shell to apply the new user group permissions."
+success -e "\t Which were: sudo usermod -aG docker \$USER"
